+++ src/main/java/com/storder/order/notice/controller/NoticeController.java
@@ -1,8 +1,8 @@
 package com.storder.order.notice.controller;
 
 import com.storder.order.global.dto.ApiResponse;
-import com.storder.order.notice.dto.response.NoticeDetailResponse;
-import com.storder.order.notice.dto.response.NoticeListResponse;
+import com.storder.order.notice.dto.user.NoticeDetailResponse;
+import com.storder.order.notice.dto.user.NoticeListResponse;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.Parameter;
 import io.swagger.v3.oas.annotations.tags.Tag;
@@ -12,8 +12,6 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
-
-import java.util.List;
 
 @RequiredArgsConstructor
 @RestController
