+++ src/main/java/com/storder/order/notice/dto/user/NoticeDetailResponse.java
@@ -1,15 +1,14 @@
-package com.storder.order.notice.dto.response;
+package com.storder.order.notice.dto.user;
 
 import io.swagger.v3.oas.annotations.media.Schema;
-import lombok.Builder;
-import lombok.Data;
+import lombok.AllArgsConstructor;
+import lombok.Getter;
 
 import java.time.LocalDateTime;
 
-@Data
-@Schema(description = "공지사항 목록 조회 응답 DTO")
-@Builder
-public class NoticeListResponse {
+@Getter
+@AllArgsConstructor
+public class NoticeDetailResponse {
 
     @Schema(description = "공지사항 ID", example = "1")
     private Long noticeId;
@@ -25,6 +24,12 @@ public class NoticeListResponse {
 
     @Schema(description = "공지사항 카테고리", example = "신메뉴 출시")
     private String category;
+
+    @Schema(description = "관련 이미지", example = "/images/notice1.png")
+    private String image;
+
+    @Schema(description = "내용", example = "사나이의 맛!! 제육덮밥 출시\n2025년 1월부터 요청이 많았던 제육덮밥을 드디어 출시합니다!!\n바비든든만의 비법소스로 메콤한 돼지고기 제육덮밥을 즐겨보세요")
+    private String content;
 
     @Schema(description = "생성일시", example = "2021-07-01T00:00:00")
     private LocalDateTime createdAt;
